requirements.txt
django==2.1.3
psycopg2==2.7.6.1
django-crispy-forms==1.7.2